+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 2.28.0 (??)
 - Feature: new cosmetic finder to fix cursive within links
+- Save more details about a reviewed replacement. Save if the replacement has been modified or not during the review, along with the timestamp and the revisions before and after in case we need to revert it.
 - Fix: save the timestamp returned by Wikipedia API instead of the local date
+- Fix: some finders stop finding after the first false positive
 
 ## 2.27.1 (23 Nov 2023)
 - Fix: ignore false degrees where the degree letter belongs to a word and therefore are likely to be ordinals